+++ src/ch/epfl/leb/alica/controllers/ControllerFactory.java
@@ -20,11 +20,24 @@
 package ch.epfl.leb.alica.controllers;
 
 import ch.epfl.leb.alica.AbstractFactory;
+import ch.epfl.leb.alica.AlicaLogger;
 import ch.epfl.leb.alica.Controller;
 import ch.epfl.leb.alica.controllers.inverter.InverterSetupPanel;
 import ch.epfl.leb.alica.controllers.manual.ManualSetupPanel;
 import ch.epfl.leb.alica.controllers.pi.PI_SetupPanel;
 import ch.epfl.leb.alica.controllers.selftuningpi.SelfTuningSetupPanel;
+import ij.io.PluginClassLoader;
+import java.io.File;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.net.URLClassLoader;
+import java.util.ArrayList;
+import java.util.Enumeration;
+import java.util.jar.JarEntry;
+import java.util.jar.JarFile;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  * Controller Factory
@@ -44,9 +57,11 @@ public class ControllerFactory extends AbstractFactory<ControllerSetupPanel>{
         addSetupPanel("Manual", new ManualSetupPanel());
         addSetupPanel("Inverter", new InverterSetupPanel());
         addSetupPanel("Self-tuning PI", new SelfTuningSetupPanel());
-        
+        for (ControllerSetupPanel sp: ControllerSetupPanelLoader.getControllerSetupPanels()) {
+            addSetupPanel(sp.getName(), sp);
+        }
         // set up default choice
-        selectProduct("PI");
+        selectProduct("Self-tuning PI");
     }
     
     /**
@@ -71,5 +86,80 @@ public class ControllerFactory extends AbstractFactory<ControllerSetupPanel>{
      */
     public Controller build() {
         return getSelectedSetupPanel().initController(max_controller_output, tick_rate_ms);
+    }
+}
+
+
+class ControllerSetupPanelLoader {
+    // I use this just to print jars in mmplugins dir, whatever
+    private static PluginClassLoader class_loader = new PluginClassLoader("mmplugins/");
+    
+    /**
+     * Dynamically loads ControllerSetupPanels from mmplugins. The jar filename MUST
+     * begin with "ALICA_" for the jar to be recognized.
+     * @return list of loaded setup panels
+     */
+    public static ArrayList<ControllerSetupPanel> getControllerSetupPanels() {
+        
+        ArrayList<ControllerSetupPanel> retval = new ArrayList<ControllerSetupPanel>();
+        
+        // print urls of files in mmplugins folder
+        for (URL u: class_loader.getURLs()) {
+            // if it doesnt match desired filename, skip it
+            if (!u.toString().toUpperCase().contains("ALICA_")) {
+                continue;
+            } else {
+                AlicaLogger.getInstance().logMessage("Loading ALICA addons from:\n" + u.toString());
+            }
+            
+            // open the jar file
+            JarFile jarFile = null;
+            try {
+                jarFile = new JarFile(new File(u.toURI()));
+            } catch (IOException ex) {
+                Logger.getLogger(ControllerSetupPanelLoader.class.getName()).log(Level.SEVERE, null, ex);
+                continue;
+            } catch (URISyntaxException ex) {
+                Logger.getLogger(ControllerSetupPanelLoader.class.getName()).log(Level.SEVERE, null, ex);
+                continue;
+            }
+
+            // list all entries, and initiate classloader
+            Enumeration<JarEntry> e = jarFile.entries();
+            URL[] urls = { u };
+            // it is vital to include a parent classloader from original ALICA package
+            URLClassLoader cl = URLClassLoader.newInstance(urls, ControllerSetupPanelLoader.class.getClassLoader());
+            
+            // iterate over entries, filter out those which arent a class
+            while (e.hasMoreElements()) {
+                JarEntry je = e.nextElement();
+                if(je.isDirectory() || !je.getName().endsWith(".class")){
+                    continue;
+                }
+                // remove .class from end of string
+                String className = je.getName().substring(0,je.getName().length()-6);
+                className = className.replace('/', '.');
+                Class c = null;
+                // try to load class
+                try {
+                    c = cl.loadClass(className);
+                } catch (ClassNotFoundException ex) {
+                    Logger.getLogger(ControllerSetupPanelLoader.class.getName()).log(Level.SEVERE, null, ex);
+                    continue;
+                }
+                
+                // verify that class is subclass of ControllerSetupPanel and add it
+                if (ControllerSetupPanel.class.isAssignableFrom(c)) {
+                    try {
+                        retval.add((ControllerSetupPanel) c.newInstance());
+                    } catch (InstantiationException ex) {
+                        Logger.getLogger(ControllerSetupPanelLoader.class.getName()).log(Level.SEVERE, null, ex);
+                    } catch (IllegalAccessException ex) {
+                        Logger.getLogger(ControllerSetupPanelLoader.class.getName()).log(Level.SEVERE, null, ex);
+                    }
+                }
+            }
+        }
+        return retval;
     }
 }